exclude tests from race check

+++ cmd/xmidt-agent/config.go
@@ -279,7 +279,7 @@ type NetworkService struct {
 
 type Cloud struct {
 	PreferQuic bool
-	MaxTries   int64
+	MaxTries   int32
 }
 
 // Collect and process the configuration files and env vars and
